DPR-611 move testing of query execution into athena client test
