+++ backend/src/test/kotlin/uk/gov/justice/digital/backend/service/PreviewServiceTest.kt
@@ -2,17 +2,10 @@ package uk.gov.justice.digital.backend.service
 
 import io.mockk.every
 import io.mockk.mockk
-import org.flywaydb.core.Flyway
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import org.postgresql.ds.PGSimpleDataSource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.containers.wait.strategy.Wait
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
-import uk.gov.justice.digital.backend.client.preview.AthenaPreviewClient
+import uk.gov.justice.digital.backend.client.preview.PreviewClient
 import uk.gov.justice.digital.backend.repository.DomainRepository
 import uk.gov.justice.digital.model.Domain
 import uk.gov.justice.digital.model.Status
@@ -22,42 +15,12 @@ import uk.gov.justice.digital.test.Fixtures.domain1
 import uk.gov.justice.digital.test.Fixtures.domain2
 import uk.gov.justice.digital.test.Fixtures.table1
 
-@Testcontainers
 class PreviewServiceTest {
 
-    companion object {
-        @Container
-        val postgresContainer: PostgreSQLContainer<*> = PostgreSQLContainer("postgres:15.2")
-            .withDatabaseName("test_domains")
-            .withUsername("postgres")
-            .withPassword("postgres")
-            .waitingFor(Wait.forListeningPort())
-            .withReuse(true)
-
-        val dataSource by lazy {
-            val d = PGSimpleDataSource()
-            d.setUrl(postgresContainer.jdbcUrl)
-            d.user = "postgres"
-            d.password = "postgres"
-            d
-        }
-
-        @JvmStatic
-        @BeforeAll
-        fun applyMigrations() {
-            Flyway
-                .configure()
-                .dataSource(dataSource)
-                .locations("classpath:testdb/migration")
-                .load()
-                .migrate()
-        }
-    }
-
     private val mockRepository = mockk<DomainRepository>()
+    private val mockClient = mockk<PreviewClient>()
 
-    // TODO - move this testing into the AthenaPreviewClient
-    private val underTest by lazy { PreviewService(AthenaPreviewClient(dataSource), mockRepository) }
+    private val underTest by lazy { PreviewService(mockClient, mockRepository) }
 
     @Test
     fun `it should execute the query and return the results for a valid request`() {
@@ -65,21 +28,12 @@ class PreviewServiceTest {
             listOf(table1.copy(transform = Transform("select * from test_domain", emptyList())))
         )
 
+        val mockResult = listOf(mapOf("foo" to "bar"))
+
         every { mockRepository.getDomains(any(), any()) } answers { listOf(domainWithTestDomainQuery) }
+        every { mockClient.runQuery(any()) } answers { mockResult }
 
-        val result = underTest.preview("Test Domain", Status.DRAFT, 50)
-
-        assertEquals(5, result.size)
-
-        val expectedFirstRow = mapOf(
-            "id" to "1",
-            "name" to "Foo",
-            "street" to "Montague Road",
-            "postcode" to "SW12 2PB",
-            "items" to "5"
-        )
-
-        assertEquals(expectedFirstRow, result[0])
+        assertEquals(mockResult, underTest.preview("Test Domain", Status.DRAFT, 50))
     }
 
     @Test
